requirements.txt
amqp==5.3.1
anyio==4.10.0
appdirs==1.4.4
APScheduler==3.11.0
asgiref==3.8.1
billiard==4.2.1
celery~=5.5.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.5
defusedxml==0.7.1
distlib==0.3.9
Django~=5.1.7
django-filter~=25.1
django-mathfilters==1.0.0
djangorestframework~=3.16.0
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-yasg~=1.21.10
filelock==3.18.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
kombu==5.5.4
oauthlib==3.3.1
packaging==24.2
pillow==11.1.0
pipenv==2024.4.1
platformdirs==4.3.7
prompt_toolkit==3.0.51
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv~=1.1.1
python-telegram-bot==22.3
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
requests~=2.32.4
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
virtualenv==20.29.3
virtualenv-clone==0.5.7
wcwidth==0.2.13
zipp==3.21.0

stripe~=12.4.0